place request into separate file
